reduce tests and combine files
